requirements.txt
Pillow == 9.1.1
fastapi == 0.78.0
pydantic == 1.9.1
numpy == 1.22.4
uvicorn == 0.17.6
boto3 == 1.24.9